Сверстал страницу фармакотерапии. Изменил наименования файлов. Поправил роуты
